The viewer script will now only try and create the viewer if you have write privileges, and will default to using an existing one if it cannot.

+++ src/python/viewer/viewer.py
@@ -27,15 +27,20 @@ if __name__ == '__main__':
 
     # Copy over share/uvcmetrics/amwg_viewer
     share_directory = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(get_python_lib()))), "share", "uvcmetrics", "viewer")
-    print share_directory
-    viewer_dir = os.path.join(path, "amwg_viewer")
-    if os.path.exists(viewer_dir):
-        shutil.rmtree(viewer_dir)
-    shutil.copytree(share_directory, viewer_dir)
 
-    build_viewer(os.path.join(path, "index.json"), diag_name="UVCMetrics AMWG")
+    if os.access(path, os.W_OK):
+        viewer_dir = os.path.join(path, "amwg_viewer")
+        if os.path.exists(viewer_dir):
+            shutil.rmtree(viewer_dir)
+        shutil.copytree(share_directory, viewer_dir)
 
-    should_open = raw_input("Viewer HTML generated at %s/index.html. Would you like to open in a browser? y/[n]: " % path)
-    if should_open and should_open.lower()[0] == "y":
-        import webbrowser
-        webbrowser.open("file://" + os.path.join(path, "index.html"))
+        build_viewer(os.path.join(path, "index.json"), diag_name="UVCMetrics AMWG")
+
+    if os.path.exists(os.path.join(path, "index.html")):
+        should_open = raw_input("Viewer HTML generated at %s/index.html. Would you like to open in a browser? y/[n]: " % path)
+        if should_open and should_open.lower()[0] == "y":
+            import webbrowser
+            webbrowser.open("file://" + os.path.join(path, "index.html"))
+    else:
+        print "Failed to generate the viewer."
+        sys.exit(1)
